requirements.txt
torch~=1.8.1+cu111
timm~=0.4.12
numpy~=1.21.1
pandas~=1.3.0
wandb~=0.11.0
sklearn~=0.0
scikit-learn~=0.24.2
seaborn~=0.11.1
matplotlib~=3.4.2
albumentations~=1.0.3
pillow~=8.3.1
onnx~=1.10.1
onnxruntime~=1.8.1
tqdm~=4.61.2